Move symbolic computation & model checking classes to new package symbolic.comp.
