+++ prism/src/symbolic/comp/ProbModelChecker.java
@@ -24,7 +24,7 @@
 //	
 //==============================================================================
 
-package prism;
+package symbolic.comp;
 
 import hybrid.PrismHybrid;
 
@@ -58,7 +58,23 @@ import parser.ast.RelOp;
 import parser.type.TypeBool;
 import parser.type.TypePathBool;
 import parser.type.TypePathDouble;
-import prism.LTLModelChecker.LTLProduct;
+import prism.AccuracyFactory;
+import prism.IntegerBound;
+import symbolic.comp.LTLModelChecker.LTLProduct;
+import prism.ModelType;
+import prism.OpRelOpBound;
+import prism.OptionsIntervalIteration;
+import prism.Prism;
+import prism.PrismComponent;
+import prism.PrismException;
+import prism.PrismFileLog;
+import prism.PrismNative;
+import prism.PrismNotSupportedException;
+import prism.PrismSettings;
+import prism.PrismUtils;
+import prism.StateValues;
+import prism.StateValuesDV;
+import prism.StateValuesMTBDD;
 import sparse.PrismSparse;
 import dv.DoubleVector;
 import explicit.ExportIterations;
